Move `parse_and_validate_server_name` to stringutils
I want to add more things to `stringutils` which will use it, and don't want to
add a depenedency from `synapse.util` to `synapse.http`.
